src/requirements.txt
# Core dependencies
pandas>=2.0.0
openpyxl>=3.1.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0

# Google Cloud Storage
google-cloud-storage>=2.10.0
google-api-python-client>=2.100.0

# AI/ML
openai>=1.0.0
requests>=2.31.0

# Vector processing
qdrant-client>=1.6.0
jina>=3.20.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0